+++ libs/schematics/packages-versions.constant.ts
@@ -1,0 +1,3 @@
+export const PACKAGES_VERSIONS = {
+  '@hapify/cli': '^1.3.2',
+};