+++ libs/schematics/src/generators/eslint-config/schema.d.ts
@@ -1,5 +1,4 @@
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface EslintGeneratorSchema {
-  generatedDir?: string;
   skipInstall?: boolean;
 }