+++ libs/schematics/src/generators/traxion-workspace/generator.ts
@@ -1,14 +1,11 @@
-/* eslint-disable no-await-in-loop */
+import { formatFiles, Tree } from '@nrwl/devkit';
 
-import { formatFiles, logger, Tree } from '@nrwl/devkit';
-
-import { installPackagesTask } from '../..';
+import { getLogger, installPackagesTask } from '../../helpers';
 import eslintGenerator from '../eslint-config/generator';
 import generateWorkflow from '../github-workflows/generator';
 import initWorkspaceGenerator from '../init-workspace/generator';
 import prettierGenerator from '../prettier-config/generator';
 import {
-  addWorkspaceFiles,
   createAdminApplication,
   createAngularApplication,
   createNestjsApplication,
@@ -17,68 +14,46 @@ import {
 } from './helpers';
 import { TraxionWorkspaceGeneratorSchema } from './schema';
 
-function showSucceed(message: string | null) {
-  return () => {
-    logger.debug(`✔ ${message}`);
-  };
-}
-
 export default async function traxionWorkspaceGenerator(
   tree: Tree,
   options: TraxionWorkspaceGeneratorSchema,
 ) {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const { isVerbose } = tree as any;
-
-  const log = {
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    info: (message: string) => {
-      if (isVerbose) logger.debug(`- ${message}`);
-      return isVerbose ? showSucceed(message) : () => undefined;
-    },
-    error: logger.error,
-  };
-
-  let succeed;
-
-  succeed = log.info('Update workspace configuration');
-  await initWorkspaceGenerator(tree);
-  succeed();
+  const log = getLogger(tree);
 
   const normalizedOptions = await normalizeOptions(tree, options);
-  const { generatedDir } = normalizedOptions;
+  const { createApps, useConfigs, useGitaction } = normalizedOptions;
 
-  succeed = log.info('Initializing linter');
-  await eslintGenerator(tree, { generatedDir });
-  succeed();
+  log.info('Update workspace configuration');
+  await initWorkspaceGenerator(tree, { skipInstall: true });
 
-  succeed = log.info('Initializing prettier');
-  await prettierGenerator(tree, { format: false });
-  succeed();
+  if (useConfigs) {
+    log.info('Initializing linter');
+    await eslintGenerator(tree, { skipInstall: true });
 
-  succeed = log.info('Initializing github actions');
-  await generateWorkflow(tree, { all: true });
-  succeed();
+    log.info('Initializing prettier');
+    await prettierGenerator(tree, { format: false, skipInstall: true });
+  }
 
-  succeed = log.info('Initializing nestjs application');
-  await createNestjsApplication(tree, normalizedOptions);
-  succeed();
+  if (useGitaction) {
+    log.info('Initializing github actions');
+    await generateWorkflow(tree, { all: true });
+  }
 
-  succeed = log.info('Initializing angular application');
-  await createAngularApplication(tree, normalizedOptions);
-  succeed();
+  if (createApps) {
+    log.warn('Creating applications is disabled for now');
 
-  succeed = log.info('Initializing admin application');
-  await createAdminApplication(tree, normalizedOptions);
-  succeed();
+    log.info('Initializing nestjs application');
+    await createNestjsApplication(tree, normalizedOptions);
 
-  succeed = log.info('Initializing hapify libraries');
+    log.info('Initializing angular application');
+    await createAngularApplication(tree, normalizedOptions);
+
+    log.info('Initializing admin application');
+    await createAdminApplication(tree, normalizedOptions);
+  }
+
+  log.info('Initializing hapify libraries');
   await createTemplateLibraries(tree, normalizedOptions);
-  succeed();
-
-  succeed = log.info('Add files to the workspace');
-  await addWorkspaceFiles(tree, normalizedOptions);
-  succeed();
 
   await formatFiles(tree);