+++ libs/schematics/src/generators/traxion-workspace/helpers/create-admin-application.helper.ts
@@ -1,5 +1,6 @@
 import { Tree } from '@nrwl/devkit';
 
+import { getLogger } from '../../../helpers';
 import adminGenerator from '../../admin-app/generator';
 import { NormalizedOptions } from '../schema';
 
@@ -7,10 +8,14 @@ export async function createAdminApplication(
   tree: Tree,
   options: NormalizedOptions,
 ) {
-  const { adminName, generatedImportPath } = options;
-  await adminGenerator(tree, {
-    name: adminName,
-    reactAdminImportPath: `${generatedImportPath}react-admin`,
-    rextClientImportPath: `${generatedImportPath}rext-client`,
-  });
+  const log = getLogger(tree);
+  log.warn('Admin application schematics is disabled for now');
+
+  // TODO: after migration 11 get back the admin generator
+  // const { adminName, generatedImportPath } = options;
+  // await adminGenerator(tree, {
+  //   name: adminName,
+  //   reactAdminImportPath: `${generatedImportPath}react-admin`,
+  //   rextClientImportPath: `${generatedImportPath}rext-client`,
+  // });
 }